refactor: migrate SDK tests to external TAG instance
Move SDK integration tests from in-memory gofakes3 backend to external TAG
instance started via `make s3-test-local`. Create new tests/integration/sdk
package with TestEnvironment that connects to running TAG and manages unique
test buckets.

Changes:
- Move 34 SDK tests from sdk_test.go to tests/integration/sdk/go_sdk_test.go
- Create new TestEnvironment in tests/integration/sdk/testutil.go for external TAG
- Add TestMain in tests/integration/sdk/main_test.go with graceful skip detection
- Reduce original sdk_test.go to only TLS/chunked encoding tests (3 tests)
- Add test-sdk Makefile target to run external TAG tests
- Update CI workflow to run SDK tests against make s3-test-local
- Update .gitignore to include .env file
- Remove tests requiring mock infrastructure (request counting, error injection)

Tests use unique bucket names with prefix to allow parallel execution and
avoid conflicts. Cache tests verify behavior via repeated requests rather than
X-Cache header inspection (since SDK doesn't expose response headers directly).
