+++ .github/workflows/s3-tests.yaml
@@ -61,6 +61,12 @@ jobs:
           AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
         run: make s3-tests
 
+      - name: Run SDK tests
+        env:
+          AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
+          AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
+        run: make test-sdk
+
       - name: Cleanup
         if: always()
         run: make s3-test-local-down
